requirements.txt
Flask
Flask-Limiter
python-dotenv
appwrite
gunicorn
beautifulsoup4
onesignal-python-api